+++ lib/savon/options.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "logger"
 require "httpi"
 
@@ -268,6 +269,19 @@ module Savon
       @options[:ssl_ca_cert] = cert
     end
 
+    def ssl_ciphers(ciphers)
+      @options[:ssl_ciphers] = ciphers
+    end
+
+    # Sets the ca cert path.
+    def ssl_ca_cert_path(path)
+      @options[:ssl_ca_cert_path] = path
+    end
+
+    # Sets the ssl cert store.
+    def ssl_cert_store(store)
+      @options[:ssl_cert_store] = store
+    end
 
     # HTTP basic auth credentials.
     def basic_auth(*credentials)