+++ lib/savon/qualified_message.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "gyoku"
 
 module Savon
@@ -9,7 +10,7 @@ module Savon
     end
 
     def to_hash(hash, path)
-      return unless hash
+      return hash unless hash
       return hash.map { |value| to_hash(value, path) } if hash.is_a?(Array)
       return hash.to_s unless hash.is_a?(Hash)