+++ lib/savon/soap_fault.rb
@@ -1,5 +1,4 @@
-require "savon"
-
+# frozen_string_literal: true
 module Savon
   class SOAPFault < Error