+++ spec/savon/qualified_message_spec.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "spec_helper"
 
 module Savon
@@ -54,6 +55,39 @@ module Savon
         }
 
         good_xml = %(<ns:Foo><Cash attr1="val1">Chunky Bacon</Cash><ns:Bar attr2="val2"><Zing>pow</Zing></ns:Bar><SelfClosing/></ns:Foo>)
+
+        message = described_class.new(types, used_namespaces, key_converter)
+        resulting_hash = message.to_hash(hash, ['tns'])
+        xml = Gyoku.xml(resulting_hash, key_converter: key_converter)
+
+        expect(resulting_hash).to eq good_result
+        expect(xml).to eq good_xml
+      end
+
+      it "properly handles boolean false" do
+        used_namespaces = {
+          %w(tns Foo) => 'ns'
+        }
+
+        hash = {
+          :foo => {
+            :falsey => {
+              :@attr1 => false,
+              :content! => false
+            }
+          }
+        }
+
+        good_result = {
+          "ns:Foo" => {
+            :falsey => {
+              :@attr1 => false,
+              :content! => false
+            }
+          }
+        }
+
+        good_xml = %(<ns:Foo><Falsey attr1="false">false</Falsey></ns:Foo>)
 
         message = described_class.new(types, used_namespaces, key_converter)
         resulting_hash = message.to_hash(hash, ['tns'])