+++ spec/savon/request_spec.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "spec_helper"
 require "integration/support/server"
 
@@ -5,6 +6,7 @@ describe Savon::WSDLRequest do
 
   let(:globals)      { Savon::GlobalOptions.new }
   let(:http_request) { HTTPI::Request.new }
+  let(:ciphers)      { OpenSSL::Cipher.ciphers }
 
   def new_wsdl_request
     Savon::WSDLRequest.new(globals, http_request)
@@ -74,8 +76,8 @@ describe Savon::WSDLRequest do
 
     describe "ssl version" do
       it "is set when specified" do
-        globals.ssl_version(:SSLv3)
-        http_request.auth.ssl.expects(:ssl_version=).with(:SSLv3)
+        globals.ssl_version(:TLSv1)
+        http_request.auth.ssl.expects(:ssl_version=).with(:TLSv1)
 
         new_wsdl_request.build
       end
@@ -88,14 +90,28 @@ describe Savon::WSDLRequest do
 
     describe "ssl verify mode" do
       it "is set when specified" do
-        globals.ssl_verify_mode(:none)
-        http_request.auth.ssl.expects(:verify_mode=).with(:none)
+        globals.ssl_verify_mode(:peer)
+        http_request.auth.ssl.expects(:verify_mode=).with(:peer)
 
         new_wsdl_request.build
       end
 
       it "is not set otherwise" do
         http_request.auth.ssl.expects(:verify_mode=).never
+        new_wsdl_request.build
+      end
+    end
+
+    describe "ssl ciphers" do
+      it "is set when specified" do
+        globals.ssl_ciphers(ciphers)
+        http_request.auth.ssl.expects(:ciphers=).with(ciphers)
+
+        new_wsdl_request.build
+      end
+
+      it "is not set otherwise" do
+        http_request.auth.ssl.expects(:ciphers=).never
         new_wsdl_request.build
       end
     end
@@ -246,6 +262,7 @@ describe Savon::SOAPRequest do
 
   let(:globals)      { Savon::GlobalOptions.new }
   let(:http_request) { HTTPI::Request.new }
+  let(:ciphers)      { OpenSSL::Cipher.ciphers }
 
   def new_soap_request
     Savon::SOAPRequest.new(globals, http_request)
@@ -378,8 +395,8 @@ describe Savon::SOAPRequest do
 
     describe "ssl version" do
       it "is set when specified" do
-        globals.ssl_version(:SSLv3)
-        http_request.auth.ssl.expects(:ssl_version=).with(:SSLv3)
+        globals.ssl_version(:TLSv1)
+        http_request.auth.ssl.expects(:ssl_version=).with(:TLSv1)
 
         new_soap_request.build
       end
@@ -392,14 +409,28 @@ describe Savon::SOAPRequest do
 
     describe "ssl verify mode" do
       it "is set when specified" do
-        globals.ssl_verify_mode(:none)
-        http_request.auth.ssl.expects(:verify_mode=).with(:none)
+        globals.ssl_verify_mode(:peer)
+        http_request.auth.ssl.expects(:verify_mode=).with(:peer)
 
         new_soap_request.build
       end
 
       it "is not set otherwise" do
         http_request.auth.ssl.expects(:verify_mode=).never
+        new_soap_request.build
+      end
+    end
+
+    describe "ssl ciphers" do
+      it "is set when specified" do
+        globals.ssl_ciphers(ciphers)
+        http_request.auth.ssl.expects(:ciphers=).with(ciphers)
+
+        new_soap_request.build
+      end
+
+      it "is not set otherwise" do
+        http_request.auth.ssl.expects(:ciphers=).never
         new_soap_request.build
       end
     end