+++ spec/support/adapters.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'httpi/adapter/httpclient'
 
 # Proxy adapter. Records all requests and passes them to HTTPClient