+++ spec/support/integration.rb
@@ -1,9 +1,10 @@
+# frozen_string_literal: true
 module SpecSupport
 
   def call_and_fail_gracefully(client, *args, &block)
     client.call(*args, &block)
   rescue Savon::SOAPFault => e
-    pending e.message
+    puts e.message
   end
 
 end